InputValues/AttributeModal/MoodContext component creation. Simplified MoodForm.
