+++ client/src/components/forms/moodForm/MoodForm.js
@@ -1,0 +1,66 @@
+import React, {useContext } from "react";
+import { MoodContext } from "../../../providers/MoodContext";
+import MoodSlider from "./MoodSlider";
+import { makeStyles } from "@material-ui/core/styles";
+import Typography from "@material-ui/core/Typography";
+import FormControl from "@material-ui/core/FormControl";
+import Select from "@material-ui/core/Select";
+import InputLabel from "@material-ui/core/InputLabel";
+import MenuItem from "@material-ui/core/MenuItem";
+import FormGroup from "@material-ui/core/FormGroup";
+import Container from "@material-ui/core/Container";
+import InputValues from "./InputValues"
+import AttributeModal from './AttributeModal';
+
+const MoodForm = () => {
+  const moods = useContext(MoodContext);
+  const {
+    moodStates: {
+      sleep,
+      setSleep,
+      medication,
+      setMedication,
+      medicationDosage,
+      setMedicationDosage,
+      exercise,
+      setExercise,
+      exerciseTime,
+      setExerciseTime,
+      yoga,
+      setYoga,
+      yogaTime,
+      setYogaTime,
+      meditation,
+      setMeditation,
+      meditationTime,
+      setMeditationTime
+    }
+  } = moods;
+
+ 
+  const useStyles = makeStyles(theme => ({
+    root: {
+      width: 300 + theme.spacing(3) * 2
+    },
+    margin: {
+      height: theme.spacing(3)
+    }
+  }));
+  const classes = useStyles();
+
+  return (
+    <Container>
+      <div className={classes.root}>
+        <Typography gutterBottom>Mood</Typography>
+        <MoodSlider />
+        <FormGroup>
+          
+        </FormGroup>
+
+        <AttributeModal/>
+      </div>
+    </Container>
+  );
+};
+
+export default MoodForm;
