+++ client/src/components/shared/Tracker.js
@@ -1,5 +1,5 @@
 import React from "react";
-import MoodForm from "../forms/MoodForm";
+import MoodForm from "../forms/moodForm/MoodForm";
 
 const Tracker = () => {
   return (
